different middleware ordering
For apps running with `config.threadsafe!`, there is no `Rack::Lock`.

Also, this is the default in Rails 4:

  http://tenderlovemaking.com/2012/06/18/removing-config-threadsafe.html

My rationale for after `ActionDispatch::Static` is that it's unlkely
that any asset will be accessed with www, as this isn't the sort
of thing a user types manually. It's also typically in the same position
as "before Rack::Lock", so it's a minimal behavior change.